requirements.txt
astroid==2.4.2
autopep8==1.5.4
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
idna==2.10
isort==5.7.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysql-connector-python==8.0.23
protobuf==3.14.0
pycodestyle==2.6.0
pylint==2.6.0
requests==2.25.1
six==1.15.0
toml==0.10.1
urllib3==1.26.2
wrapt==1.12.1
